Move legacy form dropdown tests to own file
